requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
kafka-python==2.3.1
redis==5.0.8
psycopg2-binary==2.9.11
pydantic==2.7.4
requests==2.32.3
faker==26.0.0
python-dotenv==1.0.1
